Moved all the logic into new IndexHandling class
